use_case_2/metagpt/devstral_24b_small_2505_fp16/requirements.txt
django-cors-headers==3.13.0
Django==4.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1